requirements.txt
anyio==3.7.0
cfgv==3.3.1
distlib==0.3.6
exceptiongroup==1.1.1
fastapi==0.95.2
filelock==3.12.0
identify==2.5.24
idna==3.4
mysql-connector-python==8.0.33
nodeenv==1.8.0
platformdirs==3.5.1
pre-commit==3.3.2
protobuf==3.20.3
pydantic==1.10.8
PyJWT==2.7.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.2
virtualenv==20.23.0
